landing/requirements.txt
quart==0.19.4
python-dotenv==1.0.0
